Auto
added Autonomous which...works...kinda? Made some changes to PID so it
reaches levels supa fast. Changed buttons on controller so now button 10
is the floor level so totes can be picked up. Added a Trashcan level
button which is button 4.  Pick up tote button works! and so does button
12 which lowers lift until it hits the home switch on the robot then
resets the encoder.